requirements.txt
moderngl==5.8.2
pygame_ce==2.3.2
